Move signals from models.py to signals.py
