+++ backend/api/apps.py
@@ -3,3 +3,7 @@ from django.apps import AppConfig
 
 class ApiConfig(AppConfig):
     name = 'api'
+    verbose_name = 'Api'
+
+    def ready(self):
+        import api.signals
